include missing headers for proper declarations
garmin-forerunner-tools uses several standard C functions without
including the proper headers needed to pick up their declarations.  Some
of these functions return pointers.  Failure to include the headers
means the return type is assumed to be an int, which means the pointer
is implicitly cast to an int, losing data on 64-bit architectures.